refactor: scatter view for readability
