+++ frontend/src/instance-views/scatter/index.ts
